Initial breakout of scheduling e2es to help assist in both assignment and refactoring.
