+++ OWNERS_ALIASES
@@ -1,4 +1,14 @@
 aliases:
+  sig-scheduling-maintainers:
+    - davidopp
+    - timothysc
+    - wojtek-t
+  sig-scheduling:
+    - davidopp
+    - timothysc
+    - wojtek-t
+    - k82cn 
+    - jayunit100
   sig-cli-maintainers:
     - adohe
     - brendandburns
