+++ plugin/pkg/scheduler/OWNERS
@@ -6,3 +6,5 @@ reviewers:
 - davidopp
 - timothysc
 - wojtek-t
+- k82cn 
+- jayunit100
