+++ test/e2e/BUILD
@@ -66,7 +66,6 @@ go_library(
         "networking.go",
         "networking_perf.go",
         "nodeoutofdisk.go",
-        "opaque_resource.go",
         "pd.go",
         "persistent_volumes.go",
         "persistent_volumes-disruptive.go",
@@ -82,11 +81,9 @@ go_library(
         "rc.go",
         "reboot.go",
         "replica_set.go",
-        "rescheduler.go",
         "resize_nodes.go",
         "resource_quota.go",
         "restart.go",
-        "scheduler_predicates.go",
         "security_context.go",
         "service.go",
         "service_accounts.go",
@@ -147,7 +144,6 @@ go_library(
         "//pkg/util:go_default_library",
         "//pkg/util/exec:go_default_library",
         "//pkg/util/logs:go_default_library",
-        "//pkg/util/system:go_default_library",
         "//pkg/util/version:go_default_library",
         "//pkg/volume/util/volumehelper:go_default_library",
         "//plugin/pkg/admission/serviceaccount:go_default_library",
@@ -156,6 +152,7 @@ go_library(
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/e2e/generated:go_default_library",
+        "//test/e2e/scheduling:go_default_library",
         "//test/e2e/upgrades:go_default_library",
         "//test/e2e_federation:go_default_library",
         "//test/images/net/nat:go_default_library",
@@ -173,7 +170,6 @@ go_library(
         "//vendor:github.com/onsi/ginkgo/config",
         "//vendor:github.com/onsi/ginkgo/reporters",
         "//vendor:github.com/onsi/gomega",
-        "//vendor:github.com/stretchr/testify/assert",
         "//vendor:github.com/vmware/govmomi/find",
         "//vendor:github.com/vmware/govmomi/vim25/types",
         "//vendor:golang.org/x/crypto/ssh",
@@ -230,6 +226,7 @@ go_test(
         "//pkg/client/clientset_generated/clientset:go_default_library",
         "//pkg/metrics:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/scheduling:go_default_library",
         "//test/utils:go_default_library",
         "//vendor:github.com/onsi/ginkgo",
         "//vendor:github.com/onsi/gomega",
@@ -267,6 +264,7 @@ filegroup(
         "//test/e2e/framework:all-srcs",
         "//test/e2e/generated:all-srcs",
         "//test/e2e/perftype:all-srcs",
+        "//test/e2e/scheduling:all-srcs",
         "//test/e2e/testing-manifests:all-srcs",
         "//test/e2e/upgrades:all-srcs",
     ],
