+++ test/e2e/limit_range.go
@@ -54,7 +54,7 @@ var _ = framework.KubeDescribe("LimitRange", func() {
 		Expect(err).NotTo(HaveOccurred())
 
 		By("Creating a Pod with no resource requirements")
-		pod := newTestPod(f, "pod-no-resources", v1.ResourceList{}, v1.ResourceList{})
+		pod := f.NewTestPod("pod-no-resources", v1.ResourceList{}, v1.ResourceList{})
 		pod, err = f.ClientSet.Core().Pods(f.Namespace.Name).Create(pod)
 		Expect(err).NotTo(HaveOccurred())
 
@@ -71,7 +71,7 @@ var _ = framework.KubeDescribe("LimitRange", func() {
 		}
 
 		By("Creating a Pod with partial resource requirements")
-		pod = newTestPod(f, "pod-partial-resources", getResourceList("", "150Mi"), getResourceList("300m", ""))
+		pod = f.NewTestPod("pod-partial-resources", getResourceList("", "150Mi"), getResourceList("300m", ""))
 		pod, err = f.ClientSet.Core().Pods(f.Namespace.Name).Create(pod)
 		Expect(err).NotTo(HaveOccurred())
 
@@ -92,12 +92,12 @@ var _ = framework.KubeDescribe("LimitRange", func() {
 		}
 
 		By("Failing to create a Pod with less than min resources")
-		pod = newTestPod(f, podName, getResourceList("10m", "50Mi"), v1.ResourceList{})
+		pod = f.NewTestPod(podName, getResourceList("10m", "50Mi"), v1.ResourceList{})
 		pod, err = f.ClientSet.Core().Pods(f.Namespace.Name).Create(pod)
 		Expect(err).To(HaveOccurred())
 
 		By("Failing to create a Pod with more than max resources")
-		pod = newTestPod(f, podName, getResourceList("600m", "600Mi"), v1.ResourceList{})
+		pod = f.NewTestPod(podName, getResourceList("600m", "600Mi"), v1.ResourceList{})
 		pod, err = f.ClientSet.Core().Pods(f.Namespace.Name).Create(pod)
 		Expect(err).To(HaveOccurred())
 	})
@@ -161,27 +161,6 @@ func newLimitRange(name string, limitType v1.LimitType,
 					Default:              defaultLimit,
 					DefaultRequest:       defaultRequest,
 					MaxLimitRequestRatio: maxLimitRequestRatio,
-				},
-			},
-		},
-	}
-}
-
-// newTestPod returns a pod that has the specified requests and limits
-func newTestPod(f *framework.Framework, name string, requests v1.ResourceList, limits v1.ResourceList) *v1.Pod {
-	return &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Name:  "pause",
-					Image: framework.GetPauseImageName(f.ClientSet),
-					Resources: v1.ResourceRequirements{
-						Requests: requests,
-						Limits:   limits,
-					},
 				},
 			},
 		},
