+++ test/e2e/scheduling/OWNERS
@@ -1,0 +1,4 @@
+approvers:
+- sig-scheduling-maintainers
+reviewers:
+- sig-scheduling
